Fix borken native build by moving sensing.c related unit tests from lowlevel_simulation.c
